[webapp] Minor UI updates
Replaced text with icons in several buttons.
Modified map min/max zoom level values for resource explorer and process execution map views.
Allow users to view the current version of a workflow in read-only mode.
When the user selects a map feature, the tab with feature details is activated if not already active.